# packages/nexus-agent-core/capabilities/resource/openapi/control-plane/virtual-keys.yaml
...
openapi: 3.1.0
info:
    title: Nexus Control Plane Admin API — virtual-keys
    version: dev
    description: 'Hand-maintained against control-plane source: field names, types, optionality, status codes and IAM tiers mirror the handler code; enum value sets and required-field rules enforced imperatively in handlers are verified by the openapi-review skill.'
paths:
    /api/admin/virtual-keys:
        get:
            operationId: listVirtualKeys
            summary: List virtual keys
            description: |-
                Returns a paginated list of virtual keys. Supports free-text search and
                filtering by project, organization, owner, type, status and enabled flag.
                Non-super-admin callers are implicitly scoped to virtual keys they own
                (the ownerId filter is forced to the caller's key id).
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbRead)
            x-nexus-tier: auto
            parameters:
                - name: q
                  in: query
                  description: Free-text search term applied to the virtual key listing.
                  required: false
                  schema:
                    type: string
                - name: projectId
                  in: query
                  description: Filter to virtual keys bound to this project.
                  required: false
                  schema:
                    type: string
                - name: organizationId
                  in: query
                  description: Filter to virtual keys belonging to this organization.
                  required: false
                  schema:
                    type: string
                - name: ownerId
                  in: query
                  description: Filter to virtual keys owned by this principal. Ignored for non-super-admins, who are always scoped to their own keys.
                  required: false
                  schema:
                    type: string
                - name: vkType
                  in: query
                  description: Filter by virtual key type.
                  required: false
                  schema:
                    type: string
                    enum:
                        - personal
                        - application
                - name: vkStatus
                  in: query
                  description: Filter by lifecycle status.
                  required: false
                  schema:
                    type: string
                    enum:
                        - active
                        - pending
                        - rejected
                        - revoked
                        - expired
                - name: enabled
                  in: query
                  description: 'Filter by enabled flag. Only the literal strings "true" and "false" are honored; any other value is ignored.'
                  required: false
                  schema:
                    type: string
                    enum:
                        - "true"
                        - "false"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/vkstore_VirtualKey'
                                    total:
                                        type: integer
                            example:
                                data:
                                    - id: 6f1c2e8a-7b3d-4f9a-9c11-2a4b6c8d0e21
                                      name: production-app
                                      keyPrefix: nvk_a1b2c3d4
                                      projectId: 1f0b9c2d-3e4f-5061-7283-94a5b6c7d8e9
                                      enabled: true
                                      vkType: application
                                      vkStatus: active
                                      createdAt: 2026-01-15T08:30:00Z
                                      updatedAt: 2026-01-15T08:30:00Z
                                total: 1
        post:
            operationId: createVirtualKey
            summary: Create a virtual key
            description: |-
                Creates a virtual key and returns the generated secret exactly once in the
                `key` field of the response. The secret is hashed at rest and cannot be
                retrieved again — callers must capture it from this response.

                Only `name` is unconditionally required. `vkType` defaults to "application"
                when omitted, and `allowedModels` defaults to an empty list. When the
                resolved type is "application", `projectId` and `expiresAt` become required,
                `expiresAt` must be in the future (its distance is not bounded), and the key
                is created in "pending" status awaiting approval; other types are created
                "active".

                Names must be unique: application keys share one deployment-wide namespace,
                and personal keys are unique per owner. A taken name is rejected with 409
                `duplicate_name`.
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbCreate)
            x-nexus-tier: confirm
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                name:
                                    type: string
                                    description: Human-readable name for the virtual key.
                                projectId:
                                    type:
                                        - string
                                        - "null"
                                    description: Owning project. Required when vkType resolves to "application".
                                sourceApp:
                                    type:
                                        - string
                                        - "null"
                                    description: Optional label identifying the calling application.
                                enabled:
                                    type:
                                        - boolean
                                        - "null"
                                    description: Whether the key is enabled. Defaults to true when omitted.
                                rateLimitRpm:
                                    type:
                                        - integer
                                        - "null"
                                    description: Per-key request-per-minute limit. Unlimited when null.
                                compareEndpointRateLimitRpm:
                                    type:
                                        - integer
                                        - "null"
                                    description: Per-minute request limit for the compare endpoint.
                                allowedModels:
                                    description: |-
                                        Allowlist of models this key may call, as an array of
                                        {providerId, modelId} objects. An empty array (the default when
                                        omitted) allows all models. The AI Gateway matches each request
                                        against these provider + model ids; a body that is not an array
                                        of {providerId, modelId} objects (e.g. bare model-code strings)
                                        is rejected with 400 validation_error.
                                    type: array
                                    items:
                                        type: object
                                        properties:
                                            providerId:
                                                type: string
                                            modelId:
                                                type: string
                                        required:
                                            - providerId
                                            - modelId
                                vkType:
                                    type: string
                                    description: Virtual key type. Defaults to "application" when omitted.
                                    enum:
                                        - personal
                                        - application
                                expiresAt:
                                    type:
                                        - string
                                        - "null"
                                    format: date-time
                                    description: Expiry timestamp. Required for application keys and must be in the future; the distance is not bounded.
                            required:
                                - name
                        example:
                            name: production-app
                            projectId: 1f0b9c2d-3e4f-5061-7283-94a5b6c7d8e9
                            vkType: application
                            allowedModels:
                                - providerId: 1f0b9c2d-3e4f-5061-7283-94a5b6c7d8e9
                                  modelId: 7a8b9c0d-1e2f-3a4b-5c6d-7e8f90a1b2c3
                            enabled: true
                            expiresAt: 2026-04-15T00:00:00Z
            responses:
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "409":
                    description: |-
                        Conflict — the name is already taken (`error.type` is `duplicate_name`).
                        Application virtual-key names are unique across the deployment; personal
                        virtual-key names are unique per owner.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                            example:
                                error:
                                    message: An application virtual key with this name already exists
                                    type: duplicate_name
                                    code: ""
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "201":
                    description: |-
                        Created. The `key` field carries the raw secret and is returned only
                        on this response — store it now, it cannot be retrieved later.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    id:
                                        type: string
                                    name:
                                        type: string
                                    keyPrefix:
                                        type: string
                                    key:
                                        type: string
                                        description: The raw virtual key secret, shown only once on creation.
                            example:
                                id: 6f1c2e8a-7b3d-4f9a-9c11-2a4b6c8d0e21
                                name: production-app
                                keyPrefix: nvk_a1b2c3d4
                                projectId: 1f0b9c2d-3e4f-5061-7283-94a5b6c7d8e9
                                enabled: true
                                allowedModels:
                                    - providerId: 1f0b9c2d-3e4f-5061-7283-94a5b6c7d8e9
                                      modelId: 7a8b9c0d-1e2f-3a4b-5c6d-7e8f90a1b2c3
                                key: nvk_a1b2c3d4e5f60718293a4b5c6d7e8f90a1b2c3d4e5f60718293a4b5c6d7e8f90
                                createdAt: 2026-01-15T08:30:00Z
                                updatedAt: 2026-01-15T08:30:00Z
...
    /api/admin/virtual-keys/{id}/regenerate:
        post:
            operationId: regenerateVirtualKey
            summary: Regenerate a virtual key secret
            description: |-
                Rotates the secret of an existing virtual key. The previous hash is
                invalidated on the data plane and a fresh secret is returned exactly once
                in the `key` field — it cannot be retrieved again. Non-super-admin callers
                may only regenerate keys they own.
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbUpdate)
            x-nexus-tier: confirm
            parameters:
                - name: id
                  in: path
                  description: Virtual key id.
                  required: true
                  schema:
                    type: string
            responses:
                "404":
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "403":
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "200":
                    description: |-
                        OK. The `key` field carries the freshly generated secret and is
                        returned only on this response.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    id:
                                        type: string
                                    key:
                                        type: string
                                        description: The new raw virtual key secret, shown only once.
                                    keyPrefix:
                                        type: string
                                    message:
                                        type: string
                            example:
                                id: 6f1c2e8a-7b3d-4f9a-9c11-2a4b6c8d0e21
                                key: nvk_b2c3d4e5f60718293a4b5c6d7e8f90a1b2c3d4e5f60718293a4b5c6d7e8f901a
                                keyPrefix: nvk_b2c3d4e5
                                message: Key regenerated. Save this key — it will not be shown again.
...
    /api/admin/virtual-keys/{id}/renew:
        post:
            operationId: renewVirtualKey
            summary: Renew an active virtual key
            description: |-
                Extends the expiry of an active application virtual key. `expiresAt` is
                required and must be in the future; the distance is not bounded.
                The data plane cache is invalidated so the new expiry takes effect on the
                next request. Returns 404 when the key is not found, not active, or not an
                active application key.
            x-nexus-iam-action: iam.ResourceVirtualKey.Action(iam.VerbRenew)
            x-nexus-tier: confirm
            parameters:
                - name: id
                  in: path
                  description: Virtual key id.
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                expiresAt:
                                    type: string
                                    format: date-time
                                    description: New expiry. Must be in the future; the distance is not bounded.
                            required:
                                - expiresAt
                        example:
                            expiresAt: 2026-04-15T00:00:00Z
            responses:
                "400":
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "404":
                    description: Virtual key not found, not active, or not an active application key.
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "500":
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                additionalProperties: {}
                                properties:
                                    message:
                                        type: string
                                    expiresAt:
                                        type: string
                                        format: date-time
                            example:
                                message: Virtual key renewed
                                expiresAt: 2026-04-15T00:00:00Z
...